refactor(modulize): break app/core reverse dependency with IApplicationControl
- Create IApplicationControl interface (restart/shutdown) in core/.
- Move ReloadManager from core/ to :app/, inject IApplicationControl.
- Refactor Main.kt to expose IApplicationControl as Spring bean.
- Refactor ShutdownCliCommand to inject IApplicationControl instead of Main.
