+++ src/main/java/top/chiloven/lukosbot2/core/IApplicationControl.kt
@@ -1,0 +1,9 @@
+package top.chiloven.lukosbot2.core
+
+interface IApplicationControl {
+
+    fun restart()
+
+    fun shutdown()
+
+}
